Parte 13 - Segurança na exclusão

+++ delete.php
@@ -1,12 +1,16 @@
 <?php
-$id = $_GET['id'];
-include_once("connect.php");
+session_start();
+include_once("seguranca.php");
 
-$sql = "DELETE FROM musicanaescola WHERE Identificador = $id";
-//$sql .= "( Escola ='$escola', Cidade ='$cidade', Identificador= '$identificador', Telefone='$telefone')";
-$conexao->query($sql);
-$conexao->close();
-//include("index.php");
+if($_SESSION['email'] != ""){
+	$id = $_GET['id'];
+	include_once("connect.php");
+
+	$sql = "DELETE FROM musicanaescola WHERE Identificador = $id";
+	//$sql .= "( Escola ='$escola', Cidade ='$cidade', Identificador= '$identificador', Telefone='$telefone')";
+	$conexao->query($sql);
+	$conexao->close();
+	//include("index.php");
 ?>
 <!DOCTYPE html>
 <html lang="pt-br">
@@ -14,7 +18,7 @@ $conexao->close();
 		<meta charset="utf-8">
 		<meta http-equiv="X-UA-Compatible" content="IE=edge">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>CRUD com Bootstrap 3</title>		 
+		<title>Música na Escola</title>		 
 		<link href="css/bootstrap.min.css" rel="stylesheet">
 		<link href="css/style.css" rel="stylesheet">
 	</head>
@@ -32,18 +36,19 @@ $conexao->close();
 				</div>
 				<div id="navbar" class="navbar-collapse collapse">
 					<ul class="nav navbar-nav navbar-right">
-						<li><a href="index.php">Início</a></li>
-						<li><a href="#">Opções</a></li>
-						<li><a href="#">Perfil</a></li>
-						<li><a href="#">Ajuda</a></li>
+						<li><a href="administrador.php">Início</a></li>
+						<li><a href="#">Sobre</a></li>
+						<li><a href="#">Contato</a></li>
+						<li><a href="login.php">Sair</a></li>
 					</ul>
 				</div>
 			</div>
 		</nav>
 		<div id="main" class="container-fluid">
-			<h3 class="page-header">Cadastro deletado com sucesso.</h3>
+			<h3 class="page-header text-success">Cadastro deletado com sucesso.</h3>
 		</div>
 		<script src="js/jquery.js"></script>
 		<script src="js/bootstrap.min.js"></script>
 	</body>
 </html>
+<?php } ?>
